Completion of creation request and converters

+++ src/main/java/com/airport/util/converters/airplanes/ConverterRequestAirplanes.java
@@ -18,12 +18,12 @@ public abstract class ConverterRequestAirplanes<S, T> extends EntityConverter<S,
 		airplanes.setBuilt(entity.getBuilt());
 		airplanes.setFlightDuration(Long.valueOf (entity.getFlightDuration ()));
 		airplanes.setModel(entity.getModel());
-		airplanes.setSeats(Long.valueOf (entity.getSeats ()));
-		airplanes.setRow (Long.valueOf (entity.getRow ()));
-		airplanes.setComfortSeats (Long.valueOf (entity.getComfortSeats ()));
-		airplanes.setComfortRow (Long.valueOf (entity.getComfortRow ()));
-		airplanes.setBusinessRow (Long.valueOf (entity.getBusinessSeats ()));
-		airplanes.setBusinessRow (Long.valueOf (entity.getBusinessRow ()));
+		airplanes.setSeats(Integer.valueOf (entity.getSeats ()));
+		airplanes.setRow (Integer.valueOf (entity.getRow ()));
+		airplanes.setComfortSeats (Integer.valueOf (entity.getComfortSeats ()));
+		airplanes.setComfortRow (Integer.valueOf (entity.getComfortRow ()));
+		airplanes.setBusinessRow (Integer.valueOf (entity.getBusinessSeats ()));
+		airplanes.setBusinessRow (Integer.valueOf (entity.getBusinessRow ()));
 		return airplanes;
 	}
 
@@ -31,12 +31,12 @@ public abstract class ConverterRequestAirplanes<S, T> extends EntityConverter<S,
 		if (entity.getBuilt () != null) airplanes.setBuilt (entity.getBuilt ());
 		if (entity.getFlightDuration () != null) airplanes.setFlightDuration (Long.valueOf (entity.getFlightDuration ()));
 		if (entity.getModel () != null) airplanes.setModel (entity.getModel ());
-		if (entity.getSeats () != null) airplanes.setSeats (Long.valueOf(entity.getSeats()));
-		if (entity.getRow()!=null)airplanes.setRow (Long.valueOf (entity.getRow ()));
-		if (entity.getComfortSeats()!=null)airplanes.setComfortSeats (Long.valueOf (entity.getComfortSeats ()));
-		if (entity.getComfortRow()!=null)airplanes.setComfortRow (Long.valueOf (entity.getComfortRow ()));
-		if (entity.getBusinessSeats()!=null)airplanes.setBusinessRow (Long.valueOf (entity.getBusinessSeats ()));
-		if (entity.getBusinessRow()!=null)airplanes.setBusinessRow (Long.valueOf (entity.getBusinessRow ()));
+		if (entity.getSeats () != null) airplanes.setSeats (Integer.valueOf(entity.getSeats()));
+		if (entity.getRow()!=null)airplanes.setRow (Integer.valueOf (entity.getRow ()));
+		if (entity.getComfortSeats()!=null)airplanes.setComfortSeats (Integer.valueOf (entity.getComfortSeats ()));
+		if (entity.getComfortRow()!=null)airplanes.setComfortRow (Integer.valueOf (entity.getComfortRow ()));
+		if (entity.getBusinessSeats()!=null)airplanes.setBusinessRow (Integer.valueOf (entity.getBusinessSeats ()));
+		if (entity.getBusinessRow()!=null)airplanes.setBusinessRow (Integer.valueOf (entity.getBusinessRow ()));
 		return airplanes;
 	}
